requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.7
djangorestframework==3.8.2
python-decouple==3.1
pytz==2018.5
static3==0.7.0
xlrd==1.1.0
gunicorn==19.9.0
psycopg2==2.7.5
